refactor: move formatter classes to src/formatters
